refactor: improve ASR benchmark utils
